renamed examples and added endlines
